+++ lib/src/ctap2/base.dart
@@ -12,13 +12,20 @@ class Ctap2 {
   final CtapDevice device;
 
   final CoseConfiguration? configuration;
+  CoseConfiguration? get coseConfiguration => configuration;
   Ctap2._create(this.device, this.configuration);
 
   static Future<Ctap2> create(
     CtapDevice device, {
     CoseConfiguration? configuration,
+    CoseConfiguration? coseConfiguration,
   }) async {
-    final ctap2 = Ctap2._create(device, configuration);
+    if (configuration != null &&
+        coseConfiguration != null &&
+        !identical(configuration, coseConfiguration)) {
+      throw ArgumentError('Conflicting COSE configurations');
+    }
+    final ctap2 = Ctap2._create(device, configuration ?? coseConfiguration);
     final res = await ctap2.refreshInfo();
     if (res.status != 0) {
       throw Exception('GetInfo failed.');
@@ -26,6 +33,21 @@ class Ctap2 {
     ctap2._info = res.data;
     return ctap2;
   }
+
+  static List<int> makeGetInfoRequest() => GetInfoRequest().encode();
+  static AuthenticatorInfo parseGetInfoResponse(List<int> data) =>
+      AuthenticatorInfo.decode(data);
+  static List<int> makeClientPinRequest(ClientPinRequest request) =>
+      request.encode();
+  static ClientPinResponse parseClientPinResponse(List<int> data) =>
+      ClientPinResponse.decode(data);
+  static List<int> makeCredentialManagementRequest(
+    CredentialManagementRequest request,
+  ) => request.encode();
+  static CredentialManagementResponse parseCredentialManagementResponse(
+    List<int> data, {
+    CoseConfiguration? configuration,
+  }) => CredentialManagementResponse.decode(data, configuration: configuration);
 
   AuthenticatorInfo get info => _info;